refactor: extract PoolCard component + add URL hash session loader
- Extract PoolCard to components/PoolCard.tsx (App.tsx 505→431 lines)
- Add #session=<id> URL hash param to load completed sessions directly
